+++ crates/crb-superagent/src/mission/mod.rs
@@ -5,9 +5,10 @@ pub mod sync_fn;
 
 pub use runtime::RunMission;
 
-use anyhow::Result;
+use anyhow::{anyhow, Result};
 use async_trait::async_trait;
 use crb_agent::{Agent, Context};
+use std::any::type_name;
 
 pub trait Goal: Send + 'static {}
 
@@ -22,4 +23,23 @@ pub trait Mission: Agent {
 
 pub trait Observer<M: Mission>: Send {
     fn check(&mut self, goal: &M::Goal) -> Result<()>;
+}
+
+#[async_trait]
+pub trait Operable: Mission {
+    async fn operate(self) -> Result<Self::Goal>;
+}
+
+#[async_trait]
+impl<M: Mission> Operable for M
+where
+    Self::Context: Default,
+{
+    async fn operate(self) -> Result<M::Goal> {
+        let mut runtime = RunMission::new(self);
+        runtime
+            .perform()
+            .await
+            .ok_or_else(|| anyhow!("Mission {} failed", type_name::<Self>()))
+    }
 }